requirements.txt
africastalking==1.2.9
arrow==1.3.0
asgiref==3.8.1
blessed==1.21.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==5.2.1
django-allauth==0.61.1
django-js-asset==3.1.2
django-mptt==0.17.0
django-picklefield==3.3
django-redis==5.4.0
django-rq==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-env==1.0.0
python3-openid==3.2.0
redis==4.3.5
requests==2.32.3
requests-oauthlib==2.0.0
rq==2.3.3
schema==0.7.7
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20250516
urllib3==2.4.0
wcwidth==0.2.13
psycopg2-binary==2.9.9
gunicorn

